Stock Tracker
Tech Stock Tracker is a Python script designed for financial enthusiasts and investors who wish to keep track of the latest market movements in major tech stocks. It leverages the yfinance library to fetch and display real-time stock prices for a selected list of companies, such as Apple, Amazon, and Meta. The script provides key stock information, including the current price, daily percentage change, and a visual indicator (emoji) to quickly convey the direction of the change.